curl-openssl: fix optional dependency
Replace `libidn` with `libidn2` as a dependency.

In `curl-openssl`, the `install` block uses the `--with-libidn2`
option, which causes the build process to pick up, and link against,
the `libidn2` keg if it’s there.

However, unless `libidn2` is declared as a dependency, building a
`curl-openssl` bottle may result in a corrupt bottle, which links
to the `libidn2` keg no matter if it’s installed on the target system
or not. This happens whenever `libidn2` happens to be installed at
bottle build time.

Inspecting the `curl-openssl` bottle also reveals that it doesn’t link
against `libidn`, which is to be expected given we specify the
`--with-libidn2` option in the `install` block. So I presume the
`libidn` dependency can be removed.

CC: @mistydemeo